Fix three ghost-code type errors the verifier found
With the harness fixed, Verus loads, checks vstd (2058 verified, 0 errors)
and reaches this crate — where it rejects three things plain rustc cannot
see, because rustc erases exactly the code they live in:

* gcd.rs: `(0 % a) as nat` — an unsuffixed literal in spec position has no
  inferable type. Now `0nat % a`.
* round.rs: `Q { num: -MAX_MAG, .. }` in a spec fn — negating an `i64` in
  spec context yields `int`, not `i64`. Now cast explicitly.
* round.rs: `lemma_quotient_bound(q, rem, d, n, ..)` passed `i128` values to
  a proof fn taking `int`. Now cast at the call site.

Expect more of these before any genuine proof failure. VERIFICATION.md
records the lesson: `cargo build` passing says the executable code is
well-typed and says nothing about the specifications.

+++ src/round.rs
@@ -121,7 +121,7 @@ pub open spec fn round_frac(n: int, d: int, dir: Dir) -> Q {
         if n > 0 {
             Q { num: MAX_MAG, den: 1 }
         } else {
-            Q { num: -MAX_MAG, den: 1 }
+            Q { num: (-(MAX_MAG as int)) as i64, den: 1 }
         }
     } else {
         let rn = red_num(n, d);
@@ -744,7 +744,7 @@ pub fn shift_div(n: i128, d: i128, s: u32) -> (res: (i128, i128))
             lemma_pow2_mono(i as nat, s as nat);
         }
         assert(q * d + rem == n * pow2(0nat));
-        lemma_quotient_bound(q, rem, d, n, 0nat, s as nat);
+        lemma_quotient_bound(q as int, rem as int, d as int, n as int, 0nat, s as nat);
     }
     while i < s
         invariant
